Change eHeader enum to class enum. Try GLitter.so

+++ class/network/ClientTCP.hpp
@@ -120,6 +120,6 @@ std::string ClientTCP::add_prefix(eHeader header, T *element) {
 	std::string message;
 	message.append(reinterpret_cast<char *>(&header), sizeof(eHeader));
 	message.append(reinterpret_cast<char *>(element),
-				   ClientTCP::size_header[header]);
+				   ClientTCP::size_header[static_cast<int>(header)]);
 	return message;
 }
